Refactoring, removed unused classes
