--- server/serverAPI/model/internal/user/responseSchema/IUserResponse.ts
@@ -1,5 +1,0 @@
-import IBaseUser from "../IBaseUser";
-import IContactInformation from "../IContactInformation";
-import IIdentification from "../IIdentification";
-
-export default interface IUserResponse extends IBaseUser, IIdentification, IContactInformation {}
